Membuat Navigasi dengan data

+++ app/src/main/java/com/naufalmaulanaartocarpussavero607062300078/asesment2/model/Product.kt
@@ -5,7 +5,8 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "products")
 data class Product(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    @PrimaryKey(autoGenerate = true)
+    val id: Long = 0L,
     val name: String,
     val descProduct: String,
     val price: Double
